requirements.txt
flask
python-dotenv
cohere